+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 2.2.5
 ------
 - CO-784 Added support for product SEO attributes title, keywords, description
+- CO-785 Added support for image alt attribute
 
 2.2.4.3
 ------